feat: Sửa lỗi ko build đc do chưa khai báo di cho verify email view model - Xoá otp repository
