+++ app/src/main/java/com/pwhs/quickmem/data/remote/repository/AuthRepositoryImpl.kt
@@ -6,7 +6,9 @@ import com.pwhs.quickmem.data.mapper.auth.toModel
 import com.pwhs.quickmem.data.remote.ApiService
 import com.pwhs.quickmem.domain.model.auth.AuthResponseModel
 import com.pwhs.quickmem.domain.model.auth.LoginRequestModel
+import com.pwhs.quickmem.domain.model.auth.OtpResponseModel
 import com.pwhs.quickmem.domain.model.auth.SignupRequestModel
+import com.pwhs.quickmem.domain.model.auth.VerifyEmailResponseModel
 import com.pwhs.quickmem.domain.repository.AuthRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
@@ -37,6 +39,21 @@ class AuthRepositoryImpl @Inject constructor(
                 emit(Resources.Loading())
                 val params = signUpRequestModel.toDto()
                 val response = apiService.signUp(params)
+                emit(Resources.Success(response.toModel()))
+            } catch (e: Exception) {
+                emit(Resources.Error(e.toString()))
+            }
+        }
+    }
+
+    override suspend fun verifyEmail(
+        verifyEmailResponseModel: VerifyEmailResponseModel
+    ): Flow<Resources<OtpResponseModel>> {
+        return flow {
+            try {
+                emit(Resources.Loading())
+                val params = verifyEmailResponseModel.toDto()
+                val response = apiService.verifyEmail(params)
                 emit(Resources.Success(response.toModel()))
             } catch (e: Exception) {
                 emit(Resources.Error(e.toString()))
