--- app/src/main/java/com/pwhs/quickmem/data/remote/repository/VerifyEmailRepositoryImpl.kt
@@ -1,35 +1,0 @@
-package com.pwhs.quickmem.data.remote.repository
-
-import com.pwhs.quickmem.core.utils.Resources
-import com.pwhs.quickmem.data.mapper.auth.toDto
-import com.pwhs.quickmem.data.mapper.auth.toModel
-import com.pwhs.quickmem.data.remote.ApiService
-import com.pwhs.quickmem.domain.model.auth.AuthResponseModel
-import com.pwhs.quickmem.domain.model.auth.OtpResponseModel
-import com.pwhs.quickmem.domain.model.auth.SignupRequestModel
-import com.pwhs.quickmem.domain.model.auth.VerifyEmailResponseModel
-import com.pwhs.quickmem.domain.repository.AuthRepository
-import com.pwhs.quickmem.domain.repository.OtpRepository
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-import javax.inject.Inject
-
-class VerifyEmailRepositoryImpl @Inject constructor (
-    private val apiService: ApiService
-) : OtpRepository {
-    override suspend fun verifyEmail(
-        verifyEmailResponseModel: VerifyEmailResponseModel
-    ): Flow<Resources<OtpResponseModel>> {
-        return flow {
-            try {
-                emit(Resources.Loading())
-                val params = verifyEmailResponseModel.toDto()
-                val response = apiService.verifyEmail(params)
-                emit(Resources.Success(response.toModel()))
-            } catch (e: Exception) {
-                emit(Resources.Error(e.toString()))
-            }
-        }
-    }
-
-}
