+++ app/src/main/java/com/pwhs/quickmem/domain/repository/AuthRepository.kt
@@ -2,7 +2,9 @@ package com.pwhs.quickmem.domain.repository
 
 import com.pwhs.quickmem.core.utils.Resources
 import com.pwhs.quickmem.domain.model.auth.AuthResponseModel
+import com.pwhs.quickmem.domain.model.auth.OtpResponseModel
 import com.pwhs.quickmem.domain.model.auth.SignupRequestModel
+import com.pwhs.quickmem.domain.model.auth.VerifyEmailResponseModel
 import kotlinx.coroutines.flow.Flow
 
 interface AuthRepository {
@@ -10,5 +12,8 @@ interface AuthRepository {
     suspend fun signup(
         signUpRequestModel: SignupRequestModel
     ): Flow<Resources<AuthResponseModel>>
+    suspend fun verifyEmail(
+        verifyEmailResponseModel: VerifyEmailResponseModel
+    ): Flow<Resources<OtpResponseModel>>
 
 }
