--- app/src/main/java/com/pwhs/quickmem/domain/repository/OtpRepository.kt
@@ -1,14 +1,0 @@
-package com.pwhs.quickmem.domain.repository
-
-import com.pwhs.quickmem.core.utils.Resources
-import com.pwhs.quickmem.domain.model.auth.AuthResponseModel
-import com.pwhs.quickmem.domain.model.auth.OtpResponseModel
-import com.pwhs.quickmem.domain.model.auth.SignupRequestModel
-import com.pwhs.quickmem.domain.model.auth.VerifyEmailResponseModel
-import kotlinx.coroutines.flow.Flow
-
-interface OtpRepository {
-    suspend fun verifyEmail(
-        verifyEmailResponseModel: VerifyEmailResponseModel
-    ): Flow<Resources<OtpResponseModel>>
-}
