+++ app/src/main/java/com/pwhs/quickmem/presentation/auth/verify_email/VerifyEmailUiAction.kt
@@ -1,7 +1,5 @@
 package com.pwhs.quickmem.presentation.auth.verify_email
 
-import com.pwhs.quickmem.core.data.UserRole
-
 sealed class VerifyEmailUiAction {
     data class OtpChange(val otp: String) : VerifyEmailUiAction()
     data class EmailChange(val email: String) : VerifyEmailUiAction()
