+++ app/src/main/java/com/pwhs/quickmem/presentation/auth/verify_email/VerifyEmailViewModel.kt
@@ -2,28 +2,23 @@ package com.pwhs.quickmem.presentation.auth.verify_email
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.pwhs.quickmem.core.utils.Resources
 import com.pwhs.quickmem.domain.model.auth.VerifyEmailResponseModel
 import com.pwhs.quickmem.domain.repository.AuthRepository
-import com.pwhs.quickmem.domain.repository.OtpRepository
-import com.pwhs.quickmem.presentation.auth.signup.email.SignUpWithEmailUiEvent
-import com.pwhs.quickmem.presentation.auth.signup.email.SignUpWithEmailUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.receiveAsFlow
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import timber.log.Timber
 import javax.inject.Inject
 
 @HiltViewModel
 class VerifyEmailViewModel @Inject constructor(
-    private val otpRepository: OtpRepository,
+    private val authRepository: AuthRepository,
     application: Application
 ) : AndroidViewModel(application) {
     private val _uiState = MutableStateFlow(VerifyEmailUiState())
@@ -34,7 +29,7 @@ class VerifyEmailViewModel @Inject constructor(
 
     fun verifyEmail() {
         viewModelScope.launch {
-            var response = otpRepository.verifyEmail(
+            var response = authRepository.verifyEmail(
                 VerifyEmailResponseModel(
                     email = uiState.value.email,
                     otp = uiState.value.otp
@@ -46,9 +41,11 @@ class VerifyEmailViewModel @Inject constructor(
                     is Resources.Loading -> {
                         // Show loading
                     }
+
                     is Resources.Success -> {
                         _uiEvent.send(VerifyEmailUiEvent.VerifySuccess)
                     }
+
                     is Resources.Error -> {
                         Timber.e(resource.message)
                         _uiEvent.send(VerifyEmailUiEvent.VerifyFailure)
